added the first implementation f the birdview camera

+++ lib/utils/Debug.cpp
@@ -1,4 +1,4 @@
 #include "Debug.h"
 
 
-uint8_t debugMask = 0; // default off
+uint16_t debugMask = 0; // default off
